[css-contain][css-grid] Support size containment on grid containers
We were always sizing grid containers with "contain: size" as 0x0.
However this is wrong as discussed on the following GitHub issue:
https://github.com/w3c/csswg-drafts/issues/2804

To do this we add a RunMode to GridTrackSizingAlgorithm(),
so we can run only the initialization step instead
of the full algorithm.
Using that mode we can get the expected track sizes
for size containment.

For widths the change is pretty simple
in LayoutGrid::ComputeIntrinsicLogicalWidths().
For heights we don't have a phase to compute the intrinsic size
so apart from the changes in LayoutGrid::UpdateBlockLayout()
we need some extra checks in LayoutBox too.

There is a minor change in test contain-size-grid-002.html,
because after this patch it was not passing yet due to some overflow
that is unrelated to the purpose of the test.

BUG=855261
TEST=external/wpt/css/css-contain/contain-size-grid-002.html
TEST=external/wpt/css/css-contain/contain-size-grid-003.html

Change-Id: I4d0fd417183068518070721afde84efdbfe1fcb4